Remove benchmark. Remove PipeCopyOperation class
